init project

+++ src/main/java/com/liyuan/community/dto/QuestionDto.java
@@ -5,11 +5,11 @@ import lombok.Data;
 
 @Data
 public class QuestionDto {
-    private int id;
+    private long id;
     private String title;
     private String description;
     private String tag;
-    private int creator;
+    private long creator;
     private int viewCount;
     private int commentCount;
     private int likeCount;
